refactor(http-cache): simplify test suite and extract shared fixtures
Reduced test code by 42% while maintaining coverage:
- Extracted MockAsyncStorage and MockTransport to shared conftest.py
  (~200 lines eliminated)
- Simplified test_models.py from 19 tests to 4 focused tests
  (326→96 lines, -71%)
- Removed test_cache_headers.py (low value: 2/3 xfail tests)
- Fixed imports in test_transports.py (hishel.httpx.AsyncCacheTransport)
